backend/requirements.txt
django
djangorestframework
django-cors-headers
django-dotenv
google-genai
groq
python-dotenv
pymongo
gspread
google-auth
celery
redis
pymupdf
reportlab
pypdf
gunicorn
whitenoise
axios
